requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
google-generativeai==0.8.6
python-dotenv>=1.0.0
python-multipart>=0.0.9

